Renamed file from interface.ts to interfaces.ts
