+++ src/routes/auth/interfaces.ts
