--- src/routes/cars/interface.ts
@@ -1,10 +1,0 @@
-export interface CreateCarDto {
-  licensePlate: string;
-  model: string;
-}
-export interface CreateCarModelDto {
-  licensePlate: string;
-  model: string;
-  imageFilename: string;
-}
-export interface UpdateCarDto {}
